requirements.txt
Flask 
pillow 
reportlab